backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

VAPI_API_KEY = os.getenv("VAPI_API_KEY", "")
VAPI_PHONE_NUMBER_ID = os.getenv("VAPI_PHONE_NUMBER_ID", "")
SERVER_BASE_URL = os.getenv("SERVER_BASE_URL", "http://localhost:8000")
