backend/requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
sqlalchemy==2.0.46
pydantic==2.12.5
httpx==0.28.1
python-dotenv==1.1.0
